updates to denoising image progress: - save the input / noise / timestep across experiments to get comparable results and take fewer columns in the image. - renamed lots of variables and methods to be intuitive/consistent
fixed bug in train_denoise which was causing all the experiments to be run with the same last parameters. this is the same bug i had before where i wasn't properly capturing args in a lambda expression.

misc cmdline handling updates. added --no_log to ImageTrainerConfig, to disable checkpoints and tensorboard.